Breaking changes

+++ Assets/Component_Core/Scripts/ScenarioManager.cs
@@ -1,13 +1,10 @@
 ﻿using System.Collections.Generic;
 using System.IO;
-using System.Security.Cryptography.X509Certificates;
 using Newtonsoft.Json;
-using Oculus.Platform;
 using UnityEngine;
-using Application = UnityEngine.Application;
 
-public class ScenarioManager : MonoBehaviour
-{
+public class ScenarioManager : MonoBehaviour {
+    public static string OverwriteQNDataFolderName = "QN_Updates";
     public TextAsset QuestionairToAsk;
 
     public string conditionName;
@@ -16,12 +13,6 @@ public class ScenarioManager : MonoBehaviour
     public GameObject QuestionairPrefab;
 
     public SceneField VisualSceneToUse;
-    QNSelectionManager qnmanager;
-    public bool ready{ get; private set; } // property
-    private Dictionary<ParticipantOrder, Pose> MySpawnPositions;
-
-
-    private Transform MyLocalClient;
 
     public Dictionary<ParticipantOrder, ConnectionAndSpawning.ParticipantObjectSpawnType> OrderToType;
 
@@ -37,9 +28,13 @@ public class ScenarioManager : MonoBehaviour
     public List<CameraSetupXC> CameraSetups;
 
 
-  
+    private Transform MyLocalClient;
+    private Dictionary<ParticipantOrder, Pose> MySpawnPositions;
+    private QNSelectionManager qnmanager;
+    public bool ready { get; private set; } // property
 
-    void Start(){
+
+    private void Start() {
         qnmanager = Instantiate(
             QuestionairPrefab).GetComponent<QNSelectionManager>();
         qnmanager.gameObject.SetActive(false);
@@ -47,43 +42,43 @@ public class ScenarioManager : MonoBehaviour
         ready = true;
 
         if (ConnectionAndSpawning.Singleton.ServerisRunning ||
-            ConnectionAndSpawning.Singleton.ServerState == ActionState.RERUN){ }
+            ConnectionAndSpawning.Singleton.ServerState == ActionState.RERUN) {
+        }
     }
 
 
-    void Update(){ }
+    private void Update() {
+    }
 
-    public bool GetStartPose(ParticipantOrder participantOrder, out Pose pose, out ConnectionAndSpawning.ParticipantObjectSpawnType spawnType){
+    public bool GetStartPose(ParticipantOrder participantOrder, out Pose pose,
+        out ConnectionAndSpawning.ParticipantObjectSpawnType spawnType) {
         GetSpawnPoints();
 
-        if (MySpawnPositions != null){
-            if(MySpawnPositions.Count > 0 && MySpawnPositions.ContainsKey(participantOrder)){
+        if (MySpawnPositions != null) {
+            if (MySpawnPositions.Count > 0 && MySpawnPositions.ContainsKey(participantOrder)) {
                 pose = MySpawnPositions[participantOrder];
                 spawnType = OrderToType[participantOrder];
                 return true;
             }
-             else{
-                Debug.Log("Did not find an assigned spawn point");
-                pose = new Pose();
-                spawnType = new ConnectionAndSpawning.ParticipantObjectSpawnType();
-                return false;
-            }
-        }
-        else{
-                Debug.Log("SpawnPoint is null");
-                pose = new Pose();
-                spawnType = new ConnectionAndSpawning.ParticipantObjectSpawnType();
-                return false;
+
+            Debug.Log("Did not find an assigned spawn point");
+            pose = new Pose();
+            spawnType = new ConnectionAndSpawning.ParticipantObjectSpawnType();
+            return false;
         }
 
+        Debug.Log("SpawnPoint is null");
+        pose = new Pose();
+        spawnType = new ConnectionAndSpawning.ParticipantObjectSpawnType();
+        return false;
     }
 
 
-    private void GetSpawnPoints(){
-        if (MySpawnPositions == null || MySpawnPositions.Count == 0){
+    private void GetSpawnPoints() {
+        if (MySpawnPositions == null || MySpawnPositions.Count == 0) {
             MySpawnPositions = new Dictionary<ParticipantOrder, Pose>();
             OrderToType = new Dictionary<ParticipantOrder, ConnectionAndSpawning.ParticipantObjectSpawnType>();
-            foreach (SpawnPosition sp in GetComponentsInChildren<SpawnPosition>()){
+            foreach (var sp in GetComponentsInChildren<SpawnPosition>()) {
                 var transform1 = sp.transform;
                 MySpawnPositions[sp.StartingId] = new Pose(transform1.position, transform1.rotation);
 
@@ -94,74 +89,64 @@ public class ScenarioManager : MonoBehaviour
     }
 
 
-    public void RunQuestionairNow(Transform MyLocalClient_){
+    public void RunQuestionairNow(Transform MyLocalClient_) {
         MyLocalClient = MyLocalClient_;
-        Transform MyCar = MyLocalClient.GetComponent<ParticipantInputCapture>().GetMyCar();
+        var MyCar = MyLocalClient.GetComponent<ParticipantInputCapture>().GetMyCar();
         MyLocalClient.GetComponent<ParticipantInputCapture>().NewScenario();
 
         qnmanager.gameObject.SetActive(true);
         qnmanager.transform.localScale *= 0.1f;
         qnmanager.setRelativePosition(MyCar, -0.38f, 1.14f, 0.6f);
-        if (QuestionairToAsk != null){
+        if (QuestionairToAsk != null) {
             Debug.Log("about to setup QN");
             qnmanager.startAskingTheQuestionairs(MyLocalClient, conditionName, ConnectionAndSpawning.Singleton.lang);
             Debug.Log("Started asking questions!");
         }
 
-        foreach (QnCaptureScreenShot screenShot in FindObjectsOfType<QnCaptureScreenShot>()){
-            if (screenShot.ContainsPO(ConnectionAndSpawning.Singleton.ParticipantOrder)){
-                if (screenShot.triggered){
+        foreach (var screenShot in FindObjectsOfType<QnCaptureScreenShot>())
+            if (screenShot.ContainsPO(ConnectionAndSpawning.Singleton.ParticipantOrder))
+                if (screenShot.triggered) {
                     qnmanager.AddImage(screenShot.GetTexture());
                     MyLocalClient.GetComponent<ParticipantInputCapture>()
                         .InitiateImageTransfere(screenShot.GetTexture().EncodeToJPG(50));
                     break;
                 }
-            }
-        }
     }
 
 
-
-    public Dictionary<ParticipantOrder, GpsController.Direction> GetStartingPositions(){
-        return new Dictionary<ParticipantOrder, GpsController.Direction>(){
-            {ParticipantOrder.A, StartingDirectionParticipantA},
-            {ParticipantOrder.B, StartingDirectionParticipantB},
-            {ParticipantOrder.C, StartingDirectionParticipantC},
-            {ParticipantOrder.D, StartingDirectionParticipantD},
-            {ParticipantOrder.E, StartingDirectionParticipantE},
-            {ParticipantOrder.F, StartingDirectionParticipantF}
+    public Dictionary<ParticipantOrder, GpsController.Direction> GetStartingPositions() {
+        return new Dictionary<ParticipantOrder, GpsController.Direction> {
+            { ParticipantOrder.A, StartingDirectionParticipantA },
+            { ParticipantOrder.B, StartingDirectionParticipantB },
+            { ParticipantOrder.C, StartingDirectionParticipantC },
+            { ParticipantOrder.D, StartingDirectionParticipantD },
+            { ParticipantOrder.E, StartingDirectionParticipantE },
+            { ParticipantOrder.F, StartingDirectionParticipantF }
         };
     }
 
-    public TextAsset GetQuestionFile(){
+    public TextAsset GetQuestionFile() {
         return QuestionairToAsk;
     }
 
-    public List<QuestionnaireQuestion> ReadString(string asset){
+    public List<QuestionnaireQuestion> ReadString(string asset) {
         return JsonConvert.DeserializeObject<List<QuestionnaireQuestion>>(asset);
     }
 
-
-    public static string OverwriteQNDataFolderName = "QN_Updates";
-
-    public List<QuestionnaireQuestion> GetQuestionObject(){
-        string p = Path.Combine(Application.persistentDataPath, OverwriteQNDataFolderName);
-        string file = Path.Combine(Application.persistentDataPath, OverwriteQNDataFolderName,
+    public List<QuestionnaireQuestion> GetQuestionObject() {
+        var p = Path.Combine(Application.persistentDataPath, OverwriteQNDataFolderName);
+        var file = Path.Combine(Application.persistentDataPath, OverwriteQNDataFolderName,
             QuestionairToAsk.name + ".json");
         Debug.Log(file);
-        if (!Directory.Exists(p)){
-            Directory.CreateDirectory(p);
-        }
+        if (!Directory.Exists(p)) Directory.CreateDirectory(p);
 
-        if (!File.Exists(file)){
-            File.WriteAllText(file, QuestionairToAsk.text);
-        }
+        if (!File.Exists(file)) File.WriteAllText(file, QuestionairToAsk.text);
 
 
-        List<QuestionnaireQuestion> outval = ReadString(File.ReadAllText(file));
+        var outval = ReadString(File.ReadAllText(file));
 
-        int startCounter = 0;
-        foreach (QuestionnaireQuestion q in outval){
+        var startCounter = 0;
+        foreach (var q in outval) {
             q.setInternalID(startCounter);
             startCounter++;
         }
